refactor(common)!: move metrics config classes to org.apache.hudi.common.config.metrics
Relocates the seven metrics config classes (HoodieMetricsConfig and the
CloudWatch/Datadog/Graphite/Jmx/M3/Prometheus variants) from
org.apache.hudi.config.metrics to org.apache.hudi.common.config.metrics,
joining HoodieCommonConfig and HoodieMetadataConfig in the hudi-common
config namespace. The old package was a leftover from the classes'
original hudi-client-common home; the new location also avoids adding
a split package under hudi-io's org.apache.hudi.common.metrics.

BREAKING CHANGE: source/binary break for code that imports these
classes to build configs programmatically (e.g.
HoodieWriteConfig.Builder#withMetricsConfig) - fails loudly with
NoClassDefFoundError, fixed by an import change. Regular usage is
unaffected: all hoodie.metrics.* config keys are unchanged, reporter
selection is enum-based (no class names in config values), and custom
reporters via hoodie.metrics.reporter.class are instantiated with a
(Properties, MetricRegistry) constructor that never exposes these
types. No compatibility stub: inherited static builder methods would
fail exact-descriptor resolution on the nested Builder types anyway.
Note: mixing a new engine bundle with an older hudi-aws-bundle fails
at metrics init for CloudWatch (reflective constructor lookup); bundle
versions must match, as already required.

Verified across hudi-common, client-common, spark/java clients,
hudi-spark, utilities, sync-common, hive-sync and hudi-aws with the
metrics config/reporter test suites.
