+++ hudi-client/hudi-spark-client/src/test/java/org/apache/hudi/client/TestSparkRDDWriteClient.java
@@ -351,7 +351,7 @@ class TestSparkRDDWriteClient extends SparkClientFunctionalTestHarness {
     HoodieWriteConfig configWithIgnore = getConfigBuilder(true)
         .withPath(metaClient.getBasePath())
         .withCanIgnorePostCommitFailures(true)
-        .withMetricsConfig(org.apache.hudi.config.metrics.HoodieMetricsConfig.newBuilder()
+        .withMetricsConfig(org.apache.hudi.common.config.metrics.HoodieMetricsConfig.newBuilder()
             .on(true)
             .withReporterType("INMEMORY")
             .build())
@@ -398,7 +398,7 @@ class TestSparkRDDWriteClient extends SparkClientFunctionalTestHarness {
     HoodieWriteConfig configWithoutIgnore = getConfigBuilder(true)
         .withPath(metaClient.getBasePath())
         .withCanIgnorePostCommitFailures(false)
-        .withMetricsConfig(org.apache.hudi.config.metrics.HoodieMetricsConfig.newBuilder()
+        .withMetricsConfig(org.apache.hudi.common.config.metrics.HoodieMetricsConfig.newBuilder()
             .on(true)
             .withReporterType("INMEMORY")
             .build())
