+++ hudi-common/src/main/java/org/apache/hudi/common/config/metrics/HoodieMetricsGraphiteConfig.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.config.metrics;
+package org.apache.hudi.common.config.metrics;
 
 import org.apache.hudi.common.config.ConfigClassProperty;
 import org.apache.hudi.common.config.ConfigGroups;
@@ -28,7 +28,7 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.util.Properties;
 
-import static org.apache.hudi.config.metrics.HoodieMetricsConfig.METRIC_PREFIX;
+import static org.apache.hudi.common.config.metrics.HoodieMetricsConfig.METRIC_PREFIX;
 
 /**
  * Configs for Graphite reporter type.
