+++ hudi-common/src/main/java/org/apache/hudi/common/config/metrics/HoodieMetricsPrometheusConfig.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.config.metrics;
+package org.apache.hudi.common.config.metrics;
 
 import org.apache.hudi.common.config.ConfigClassProperty;
 import org.apache.hudi.common.config.ConfigGroups;
@@ -25,7 +25,7 @@ import org.apache.hudi.common.config.HoodieConfig;
 
 import java.util.Properties;
 
-import static org.apache.hudi.config.metrics.HoodieMetricsConfig.METRIC_PREFIX;
+import static org.apache.hudi.common.config.metrics.HoodieMetricsConfig.METRIC_PREFIX;
 
 /**
  * Configs for Prometheus/Pushgaeway reporter type.
