+++ hudi-common/src/main/java/org/apache/hudi/metrics/datadog/DatadogMetricsReporter.java
@@ -18,10 +18,10 @@
 
 package org.apache.hudi.metrics.datadog;
 
+import org.apache.hudi.common.config.metrics.HoodieMetricsConfig;
 import org.apache.hudi.common.util.Option;
 import org.apache.hudi.common.util.StringUtils;
 import org.apache.hudi.common.util.ValidationUtils;
-import org.apache.hudi.config.metrics.HoodieMetricsConfig;
 import org.apache.hudi.metrics.MetricsReporter;
 import org.apache.hudi.metrics.datadog.DatadogHttpClient.ApiSite;
