+++ hudi-common/src/main/java/org/apache/hudi/metrics/m3/M3MetricsReporter.java
@@ -18,7 +18,7 @@
 
 package org.apache.hudi.metrics.m3;
 
-import org.apache.hudi.config.metrics.HoodieMetricsConfig;
+import org.apache.hudi.common.config.metrics.HoodieMetricsConfig;
 import org.apache.hudi.metrics.MetricsReporter;
 
 import com.codahale.metrics.MetricRegistry;
