+++ hudi-common/src/test/java/org/apache/hudi/common/config/metrics/TestHoodieMetricsConfig.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.config.metrics;
+package org.apache.hudi.common.config.metrics;
 
 import org.junit.jupiter.api.Test;
